cpuidle-8x60: Add cpuidle driver for 8x60
This is a transitional version of cpuidle and supports only 3 states.
All of them call wfi but have different latencies and threshold.

A more functional cpuidle that integrates with the power management
features of 8660 is being worked on. This driver only serves the purpose of
enabling power tools such as powertop.

Change-Id: I705267e40cb7c74cd656a8247c2fec99734f52d3
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>