msm: 7x27: Flush AXI Bus buffer in msm_arch_idle()
On 7x27 the AXI bus buffer should also be flushed whenever a
cache maintenance operation is being performed. This flush is done
by writing to a strongly ordered memory location. Add this operation
to msm_arch_idle() and also take care of preserving the state of
exception flags and the link register.

Change-Id: Ib49081fa3140f13cf9bf4ce1fa2226153f3b9da3
CRs-fixed: 238082
Signed-off-by: Naveen Ramaraj <nramaraj@codeaurora.org>