regulator: msm: Fix regulator registration bug
regulator_register returns a handle that is used by the regulator
driver in the enable/disable functions of the regulator driver.
since regulator_register may call enable/disable on the regulator
driver, there's a chance that the driver may dereference the null
handle. In this patch the enable/disable functions use the handle
passed in by the regulator core rather than the handle returned
by regulator_register.

Change-Id: Ie35b5eef217aad16dd71e755c1c7a447a482c1a3
Signed-off-by: Bobby Crabtree <bobbyc@codeaurora.org>