msm: Add framebuffer read and write support in rotator driver
Rotator driver can only read or write to PMEM.
To read or write to framebuffer, we change MDP
driver to provide framebuffer physical address
information. This change is required
so that we avoid copy of framebuffer in PMEM
for rotation. The rotated buffer is then used
to push on overlay engine.

Change-Id: I17322fa77bb05f36f70ac0c225299614dd607fe9
Signed-off-by: Omprakash Dhyade <odhyade@codeaurora.org>