msm: adsp: qdsp5v2: Increase retry duration to 5msecs.
When DSP is not ready to accept a command, ADSP driver retries
for 40 times with 50 usecs delay each time which amounts to
2 msec in total. But when DSP is busy processing commands it may
take more than 2 msec to respond to requests from application
processor.

Hence updating the number of trials to 100 with 50 usecs delay each
time thus increasing the maximum delay to 5 msecs.

CRs-fixed:242965
Change-Id: I80eafe06d8b2188214e3cc81d144e2a51d470eb7
Signed-off-by: Suresh Vankadara <svankada@qualcomm.com>