mmc: ADQ: Fix resume issues with slow SD cards.
This was missed entirely because the phone I was testing with it had a new, good quality SD,
and had no problems resuming from suspend. With cheap cards, it failed almost 100% of the
time on resume:

[  120.073158] mmc0: Command CRC error
[  120.093044] mmc0: error -110 during resume (card was removed?)
[  120.093186] mmc0: card 0260 removed

and messed up the IRQs to a point where the whole ARM9 crashed. The original mmc code from
FIH had the host resume/suspend almost entirely commented out, doing the same thing here
apparently solves the problem