rtc: ADQ: Update msm_rtc to standard version
The standard RTC doesn't quite work with the ADQ's radio, so it needed some adjustments. But it now includes the timeremote functionality expected by the RPC layer, and this will hopefully be sufficient to solve the clock stability issues I've been having with suspension and get rid of the "get_smem_clock: invalid start state" errors.